refactor: domain, view, controller 폴더로 분리
