+++ __tests__/RacingGame.test.js
@@ -1,6 +1,6 @@
-import Car from '../src/Car';
-import CarList from '../src/CarList';
-import RacingGame from '../src/RacingGame';
+import Car from '../src/domain/Car';
+import CarList from '../src/domain/CarList';
+import RacingGame from '../src/domain/RacingGame';
 
 describe('RacingGame 테스트', () => {
   const carNames = ['자동차1', '자동차2', '자동차3'];
