+++ src/View/GameInput/index.js
