+++ src/View/GameViewer/index.js
